extractor/requirements.txt
pypcap
dpkt
keyboard
pandas
numpy
pickle